refactor: move DocumentArray methods to separate directory re: #8884
