+++ lib/schema/index.js
@@ -11,7 +11,7 @@ exports.Number = require('./number');
 
 exports.Boolean = require('./boolean');
 
-exports.DocumentArray = require('./DocumentArray');
+exports.DocumentArray = require('./documentarray');
 
 exports.Embedded = require('./SingleNestedPath');
